refactor logic
