+++ packages/next/src/client/components/router-reducer/create-initial-router-state.ts
@@ -90,8 +90,6 @@ export function createInitialRouterState({
       onlyHashChange: false,
       hashFragment: null,
       segmentPaths: [],
-      rafId1: null,
-      rafId2: null,
     },
     canonicalUrl,
     renderedSearch: initialRenderedSearch,
